json 모듈 추가

+++ AssetProj/Include/Asset/Animation/Animation2DManager.cpp
@@ -21,8 +21,8 @@ bool CAnimation2DManager::Init()
 		SetAnimationTextureType("PlayerIdle", EAnimationTextureType::SpriteSheet);
 		SetTexture("PlayerIdle", "PlayerSprite", TEXT("Texture\\bird_4x1_798x135.png")); // 798 * 135
 
-		int sizeX = 798 / 4;
-		int sizeY = 135;
+		float sizeX = 798 / 4;
+		float sizeY = 135;
 
 		for (int i = 0; i < 2; ++i)
 		{
